Removed 'test' feature; flagged a couple of other breaking changes.
