+++ sdk/eventhubs/azure_messaging_eventhubs/CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ### Breaking Changes
 
-- `ProducerClient::submit_batch` now consumes its `batch` argument.
+- `ProducerClient::send_batch` now consumes its `batch` argument.
+- `RetryOptions::max_retries` is a `u32` not a `usize`.
 
 ### Bugs Fixed
