+++ sdk/eventhubs/azure_messaging_eventhubs/Cargo.toml
@@ -36,14 +36,14 @@ azure_core_amqp = { workspace = true, features = ["test"] }
 azure_core_test = { workspace = true, features = ["tracing"] }
 azure_identity.workspace = true
 azure_messaging_eventhubs = { path = ".", features = [
-  "test",
+  "in_memory_checkpoint_store",
 ] }
 criterion.workspace = true
 tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
 tracing-subscriber = { workspace = true, features = ["env-filter", "fmt"] }
 
 [features]
-test = []
+in_memory_checkpoint_store = []
 blob_checkpoint_store = []
 
 [[bench]]
